По замечаниям. Переименует компонент HeaderNav в Navigation и подключит его к store напрямую.
