+++ src/components/header/header.jsx
@@ -1,14 +1,10 @@
 import React from "react";
 import PropTypes from "prop-types";
-import {connect} from "react-redux";
 
-import {AuthorizationStatus} from "../../reducer/user/user.js";
-import {getAuthStatus, getAuthInfo} from "../../reducer/user/selectors.js";
-import {ActionCreator} from "../../reducer/app/app.js";
-import HeaderNav from "../header-nav/header-nav.jsx";
+import Navigation from "../navigation/navigation.jsx";
 
 const Header = (props) => {
-  const {isLogoActive, authInfo, authStatus, onChangeAuthPageState} = props;
+  const {isLogoActive} = props;
   return (
     <header className="header">
       <div className="container">
@@ -18,11 +14,7 @@ const Header = (props) => {
               <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
             </a>
           </div>
-          <HeaderNav
-            authInfo={authInfo}
-            authStatus={authStatus}
-            onChangeAuthPageState={onChangeAuthPageState}
-          />
+          <Navigation />
         </div>
       </div>
     </header>
@@ -31,29 +23,6 @@ const Header = (props) => {
 
 Header.propTypes = {
   isLogoActive: PropTypes.bool.isRequired,
-  authInfo: PropTypes.shape({
-    avatarUrl: PropTypes.string.isRequired,
-    email: PropTypes.string.isRequired,
-    id: PropTypes.any,
-    isPro: PropTypes.bool,
-    name: PropTypes.string.isRequired,
-  }).isRequired,
-  authStatus: PropTypes.oneOf([AuthorizationStatus.AUTH, AuthorizationStatus.NO_AUTH]).isRequired,
-  onChangeAuthPageState: PropTypes.func.isRequired,
 };
 
-const mapStateToProps = (state) => {
-  return {
-    authStatus: getAuthStatus(state),
-    authInfo: getAuthInfo(state),
-  };
-};
-
-const mapDispatchToProps = (dispatch) => ({
-  onChangeAuthPageState(state) {
-    dispatch(ActionCreator.changeAuthPageState(state));
-  }
-});
-
-export {Header};
-export default connect(mapStateToProps, mapDispatchToProps)(Header);
+export default Header;
