+++ src/components/navigation/navigation.jsx
@@ -1,9 +1,12 @@
 import React from "react";
 import PropTypes from "prop-types";
+import {connect} from "react-redux";
 
 import {AuthorizationStatus} from "../../reducer/user/user.js";
+import {getAuthStatus, getAuthInfo} from "../../reducer/user/selectors.js";
+import {ActionCreator} from "../../reducer/app/app.js";
 
-const HeaderNav = (props) => {
+const Navigation = (props) => {
   const {authStatus, authInfo, onChangeAuthPageState} = props;
 
   return (
@@ -28,7 +31,7 @@ const HeaderNav = (props) => {
   );
 };
 
-HeaderNav.propTypes = {
+Navigation.propTypes = {
   authStatus: PropTypes.oneOf([AuthorizationStatus.AUTH, AuthorizationStatus.NO_AUTH]).isRequired,
   authInfo: PropTypes.shape({
     avatarUrl: PropTypes.string.isRequired,
@@ -40,4 +43,18 @@ HeaderNav.propTypes = {
   onChangeAuthPageState: PropTypes.func.isRequired,
 };
 
-export default HeaderNav;
+const mapStateToProps = (state) => {
+  return {
+    authStatus: getAuthStatus(state),
+    authInfo: getAuthInfo(state),
+  };
+};
+
+const mapDispatchToProps = (dispatch) => ({
+  onChangeAuthPageState(state) {
+    dispatch(ActionCreator.changeAuthPageState(state));
+  }
+});
+
+export {Navigation};
+export default connect(mapStateToProps, mapDispatchToProps)(Navigation);
